Simplify the KCL stdlib test codegen
