+++ src/wasm-lib/derive-docs/tests/args_with_lifetime.gen
@@ -21,76 +21,14 @@ mod test_examples_someFn {
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 5)]
     async fn serial_test_example_someFn0() {
-        let user_agent = concat!(env!("CARGO_PKG_NAME"), ".rs/", env!("CARGO_PKG_VERSION"),);
-        let http_client = reqwest::Client::builder()
-            .user_agent(user_agent)
-            .timeout(std::time::Duration::from_secs(600))
-            .connect_timeout(std::time::Duration::from_secs(60));
-        let ws_client = reqwest::Client::builder()
-            .user_agent(user_agent)
-            .timeout(std::time::Duration::from_secs(600))
-            .connect_timeout(std::time::Duration::from_secs(60))
-            .connection_verbose(true)
-            .tcp_keepalive(std::time::Duration::from_secs(600))
-            .http1_only();
-        let token = std::env::var("KITTYCAD_API_TOKEN").expect("KITTYCAD_API_TOKEN not set");
-        let mut client = kittycad::Client::new_from_reqwest(token, http_client, ws_client);
-        if let Ok(addr) = std::env::var("LOCAL_ENGINE_ADDR") {
-            client.set_base_url(addr);
-        }
-
-        let tokens = crate::token::lexer("someFn()").unwrap();
-        let parser = crate::parser::Parser::new(tokens);
-        let program = parser.ast().unwrap();
-        let ctx = crate::executor::ExecutorContext::new(&client, Default::default())
-            .await
-            .unwrap();
-        ctx.run(&program, None).await.unwrap();
-        let results = program.lint_all().unwrap();
-        if !results.is_empty() {
-            panic!("Linting failed: {:?}", results);
-        }
-
-        ctx.engine
-            .send_modeling_cmd(
-                uuid::Uuid::new_v4(),
-                crate::executor::SourceRange::default(),
-                kittycad::types::ModelingCmd::ZoomToFit {
-                    object_ids: Default::default(),
-                    padding: 0.1,
-                },
-            )
-            .await
-            .unwrap();
-        let resp = ctx
-            .engine
-            .send_modeling_cmd(
-                uuid::Uuid::new_v4(),
-                crate::executor::SourceRange::default(),
-                kittycad::types::ModelingCmd::TakeSnapshot {
-                    format: kittycad::types::ImageFormat::Png,
-                },
-            )
-            .await
-            .unwrap();
-        let output_file =
-            std::env::temp_dir().join(format!("kcl_output_{}.png", uuid::Uuid::new_v4()));
-        if let kittycad::types::OkWebSocketResponseData::Modeling {
-            modeling_response: kittycad::types::OkModelingCmdResponse::TakeSnapshot { data },
-        } = &resp
-        {
-            std::fs::write(&output_file, &data.contents.0).unwrap();
-        } else {
-            panic!("Unexpected response from engine: {:?}", resp);
-        }
-
-        let actual = image::io::Reader::open(output_file)
-            .unwrap()
-            .decode()
-            .unwrap();
+        let code = "someFn()";
+        let result =
+            crate::test_server::execute_and_snapshot(code, crate::settings::types::UnitLength::Mm)
+                .await
+                .unwrap();
         twenty_twenty::assert_image(
             &format!("tests/outputs/{}.png", "serial_test_example_someFn0"),
-            &actual,
+            &result,
             0.99,
         );
     }
