+++ src/wasm-lib/kcl/src/lib.rs
@@ -26,6 +26,7 @@ pub mod lsp;
 pub mod parser;
 pub mod settings;
 pub mod std;
+#[cfg(not(target_arch = "wasm32"))]
 pub mod test_server;
 pub mod thread;
 pub mod token;
